Logical changes to forgot-password feature

+++ api/MorganizerAPI/src/main/java/com/morganizer/repository/ResetQuestionsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.morganizer.entity.ResetQuestionsEntity;
 
-public interface ResetQuestionsRepository extends JpaRepository<ResetQuestionsEntity, Integer>{
+public interface ResetQuestionsRepository extends JpaRepository<ResetQuestionsEntity, Long>{
 
 }
